fix: type any 지우기 1차
